Refactor IO into a trait
