+++ src/emulator/io/sdl.rs
@@ -1,16 +1,12 @@
 extern crate sdl2;
 
-use std::rc::Rc;
-use std::cell::RefCell;
-
 use self::sdl2::event;
 use self::sdl2::pixels;
 use self::sdl2::render;
 use self::sdl2::video;
 
 use emulator::clock;
-use emulator::io::palette::PALETTE;
-use emulator::ppu;
+use emulator::io::{EventHandler, Graphics, Input};
 
 const SCALE: u8 = 4;
 
@@ -69,18 +65,23 @@ impl IO {
         self.canvas.present();
     }
 
-    pub fn draw_screen(&mut self, pixel_data: &[u8]) {
-        let _ = self.screen_texture.update(None, pixel_data, 256 * 3);
-        let _ = self.canvas.copy(&self.screen_texture, None, None);
-    }
-
     pub fn process_event(&mut self, event: event::Event) {
         for mut handler in self.event_handlers.iter_mut() {
             handler.handle_event(&event);
         }
     }
 
-    pub fn register_event_handler(&mut self, handler: Box<dyn EventHandler>) {
+}
+
+impl Graphics for IO {
+    fn draw_screen(&mut self, pixel_data: &[u8]) {
+        let _ = self.screen_texture.update(None, pixel_data, 256 * 3);
+        let _ = self.canvas.copy(&self.screen_texture, None, None);
+    }
+}
+
+impl Input for IO {
+    fn register_event_handler(&mut self, handler: Box<dyn EventHandler>) {
         self.event_handlers.push(handler);
     }
 }
@@ -95,69 +96,3 @@ impl clock::Ticker for IO {
     }
 }
 
-pub struct Graphics {
-    io: Rc<RefCell<IO>>,
-    scanline: u32,
-    dot: u32,
-    screen_buffer: [u8; 256 * 240 * 3],
-    render_tile_grid: bool,
-}
-
-impl ppu::VideoOut for Graphics {
-    fn emit(&mut self, c: ppu::Colour) {
-        let x = self.dot;
-        let y = self.scanline;
-
-        let colour = if  self.render_tile_grid && (x % 8 == 0 || y % 8 == 0) {
-            pixels::Color::RGB(255, 0, 0)
-        } else {
-            Graphics::convert_colour(c)
-        };
-
-        self.screen_buffer[((x + y * 256) * 3) as usize] = colour.r;
-        self.screen_buffer[((x + y * 256) * 3 + 1) as usize] = colour.g;
-        self.screen_buffer[((x + y * 256) * 3 + 2) as usize] = colour.b;
-
-        self.dot = (self.dot + 1) % 256;
-        if self.dot == 0 {
-            self.scanline = (self.scanline + 1) % 240;
-            if self.scanline == 0 {
-                self.render();
-            }
-        }
-    }
-}
-
-impl Graphics {
-    pub fn new(io: Rc<RefCell<IO>>) -> Graphics {
-        Graphics {
-            io,
-            scanline: 0,
-            dot: 0,
-            screen_buffer: [0; 256 * 240 * 3],
-            render_tile_grid: false,
-        }
-    }
-
-    fn render(&mut self) {
-        self.io.borrow_mut().draw_screen(&self.screen_buffer);
-    }
-
-    fn convert_colour(c: ppu::Colour) -> pixels::Color {
-        let (r, g, b) = match PALETTE.get(c.as_byte() as usize) {
-            None => (0, 0, 0),
-            Some(colour) => *colour,
-        };
-        pixels::Color::RGB(r, g, b)
-    }
-}
-
-pub trait EventHandler {
-    fn handle_event(&mut self, event: &event::Event);
-}
-
-impl <H : EventHandler> EventHandler for Rc<RefCell<H>> {
-    fn handle_event(&mut self, event: &event::Event) {
-        self.borrow_mut().handle_event(event);
-    }
-}
